xrefer: address mentor skim-review — lazy vivisect import, optional lief, scoped --force
- factory.py: VivisectBackendFactory now lives here with a deferred VivisectBackend
  import + find_spec() availability check, so plugin startup no longer imports
  vivisect (IDA/Binja/Ghidra loads stay clean and don't fail if vivisect is broken).
  Removed the old factory from vivisect/backend.py and dropped it from __all__.
- pyproject.toml: moved lief out of core dependencies into a 'vivisect' extra —
  only the vivisect backend uses it (lazily), so other users don't install it.
- cli.py: scoped --force cleanup to the active backend's own outputs so a run
  never deletes another backend's .xrefer/report artifacts.
